refactor: IndexDataDto 사용 정리
- 일부 파일명 수정
- IndexPerformanceDto 필드 타입 변경

+++ src/main/java/com/kueennevercry/findex/dto/response/IndexPerformanceDto.java
@@ -6,8 +6,8 @@ public record IndexPerformanceDto(
     Long indexInfoId,
     String indexClassification,
     String indexName,
-    float versus,
-    float fluctuationRate,
+    BigDecimal versus,
+    BigDecimal fluctuationRate,
     BigDecimal currentPrice,
     BigDecimal beforePrice
 ) {
